requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI/LLM
langchain>=0.2.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-mcp-adapters>=0.1.0
openai>=1.0.0

# Vector Database & RAG
faiss-cpu>=1.7.4
langchain-text-splitters>=0.2.0
pypdf>=3.0.0

# GitHub Integration
requests>=2.31.0

# Google Drive (optional - for drive_loader)
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0

# Security Scanners (for local scanning)
bandit>=1.7.0
semgrep>=1.0.0
safety>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
pyyaml>=6.0.0
beautifulsoup4>=4.12.0
